feat: Add username to profiles/me endpoint and fix Django admin file upload
- Added username field to UserProfileBasicSerializer for /api/v1/profiles/me/ endpoint
- Fixed TypeError in Django admin file uploads by converting FILE_UPLOAD_MAX_MEMORY_SIZE and DATA_UPLOAD_MAX_MEMORY_SIZE to integers
- Fixed multipart parser slice indices error when uploading photos in admin
- Updated create_production_admin.py to remove decouple dependency

+++ vineyard_group_fellowship/settings/base.py
@@ -391,8 +391,9 @@ CACHES = {
 # ============================================================================
 
 # File upload settings (optimized for base64 storage)
-FILE_UPLOAD_MAX_MEMORY_SIZE = 1.5 * 1024 * 1024  # 1.5MB raw file size
-DATA_UPLOAD_MAX_MEMORY_SIZE = 1.5 * 1024 * 1024  # 1.5MB
+# Must be integers for Django's multipart parser
+FILE_UPLOAD_MAX_MEMORY_SIZE = int(1.5 * 1024 * 1024)  # 1.5MB raw file size (1572864 bytes)
+DATA_UPLOAD_MAX_MEMORY_SIZE = int(1.5 * 1024 * 1024)  # 1.5MB (1572864 bytes)
 
 # Profile photo specific settings for base64 storage
 PROFILE_PHOTO_SETTINGS = {
